STASH scene drawing WIP
About to get on a flight! Rewrite this later
